==> mRNA/1.Read_Filter_raw_counts.Rmd <==
---
title: "Read in Raw mRNA Counts, and Filter"
author: "Benji Lamp"
date: "`r Sys.Date()`"
output:
  html_document:
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
# Libraries #####################################################################
library(dplyr)
library(readxl)
library(tidyr)
library(here)
```


```{r}
# Raw Count Data From Galaxy #####################################################
file_path <- 'https://raw.githubusercontent.com/blamp23/transcriptional_dynamics_of_bov_mam/main/mRNA/raw_data/mRNA_raw_counts.tabular'

df <- read.delim(file_path, header = TRUE, row.names = 1)
df <- df[-1]

# Modify samples data frame with capitalized time points and spaces
samples <- data.frame(
  row.names = c("V 468", "V 502", "V 504", "V 507", "V 509", "V 598", "V 610", 
                "MP 468", "MP 502", "MP 504", "MP 507", "MP 509", "MP 598", "MP 610",
                "LP 468", "LP 502", "LP 504", "LP 509", "LP 507", "LP 598", "LP 610",
                "EL 468", "EL 502", "EL 504", "EL 507", "EL 509", "EL 598", "EL 610",
                "PL 468", "PL 502", "PL 504", "PL 507", "PL 509", "PL 598", "PL 610"),
  condition = c(rep("V", 7), rep("MP", 7), rep("LP", 7), rep("EL", 7), rep("PL", 7)),
  animal = c(468, 502, 504, 507, 509, 598, 610,
             468, 502, 504, 507, 509, 598, 610,
             468, 502, 504, 509, 507, 598, 610,
             468, 502, 504, 507, 509, 598, 610,
             468, 502, 504, 507, 509, 598, 610),
  batch = c(1, 1, 1, 1, 1, 1, 1, 
            0, 1, 0, 1, 0, 0, 0,
            0, 1, 0, 0, 1, 0, 0,
            0, 1, 0, 1, 0, 0, 0, 
            1, 1, 1, 1, 1, 1, 1 )
)

# Update the column names with capitalized time points and spaces
new_colnames <- c(row.names(samples))
colnames(df) <- new_colnames

# Correct swapped samples
colnames(df)[28] <- "LP 610"  # LP 610 and EL 610 are swapped samples
colnames(df)[21] <- "EL 610"

el_610 <- which(names(df) == "EL 610")
lp_610 <- which(names(df) == "LP 610")

# Create a new vector of column names with the positions of these two columns swapped
new_order <- names(df)
new_order[c(el_610, lp_610)] <- new_order[c(lp_610, el_610)]

# Reorder the columns in the DataFrame
df <- df[, new_order]

# Ensure the samples data frame matches the reordered columns
samples <- samples[colnames(df), ]

```


```{r}
# Filter Method A ##############################################################
# originaly 27607 genes
# filtered to 17344 genes
# Step 1: Splitting Data by Condition
conditions <- unique(samples$condition)
split_data <- lapply(conditions, function(cond) {
  df[, samples$condition == cond]
})

# Step 2: Define Filtering Function
count_high_expressions <- function(data) {
  apply(data, 1, function(gene) {
    sum(gene >= 3) >= 4  # Returns TRUE if at least 4 out of 7 samples have counts ≥ 3
  })
}
high_exp_counts <- lapply(split_data, count_high_expressions)

# Step 3: Aggregate Results for inclusion
gene_inclusion_filter <- rowSums(do.call(cbind, high_exp_counts)) >= 1 # If 1/5 timepoints meet crit, gene is included

# Step 4: Filter Genes
df <- df[gene_inclusion_filter, ]
```

```{r}
x <- ls()
x <- x[!x %in% c("df", "samples")]
rm(list = x)
rm(x)

save.image(here("mRNA/processed_data", "mRNA_filtered_rawcounts.RData"))
```
